+++ multibody/fem/isoparametric_element.cc
@@ -1,0 +1,1 @@
+#include "drake/multibody/fem/isoparametric_element.h"
