+++ multibody/fem/linear_simplex_element.h
@@ -4,7 +4,7 @@
 #include <utility>
 
 #include "drake/common/eigen_types.h"
-#include "drake/multibody/fixed_fem/dev/isoparametric_element.h"
+#include "drake/multibody/fem/isoparametric_element.h"
 
 namespace drake {
 namespace multibody {
