+++ multibody/fixed_fem/dev/BUILD.bazel
@@ -167,10 +167,10 @@ drake_cc_library(
         ":dynamic_elasticity_model",
         ":fem_model_base",
         ":linear_constitutive_model",
-        ":linear_simplex_element",
         ":mesh_utilities",
         "//common:essential",
         "//geometry/proximity:volume_mesh",
+        "//multibody/fem:linear_simplex_element",
         "//multibody/fem:simplex_gaussian_quadrature",
         "//multibody/plant",
     ],
@@ -278,9 +278,9 @@ drake_cc_library(
     deps = [
         ":constitutive_model",
         ":fem_element",
-        ":isoparametric_element",
         "//common:essential",
         "//common:unused",
+        "//multibody/fem:isoparametric_element",
         "//multibody/fem:quadrature",
     ],
 )
@@ -403,19 +403,6 @@ drake_cc_library(
 )
 
 drake_cc_library(
-    name = "isoparametric_element",
-    srcs = [
-        "isoparametric_element.cc",
-    ],
-    hdrs = [
-        "isoparametric_element.h",
-    ],
-    deps = [
-        "//common:essential",
-    ],
-)
-
-drake_cc_library(
     name = "linear_constitutive_model",
     hdrs = [
         "linear_constitutive_model.h",
@@ -440,20 +427,6 @@ drake_cc_library(
     deps = [
         ":deformation_gradient_data",
         "//common:essential",
-    ],
-)
-
-drake_cc_library(
-    name = "linear_simplex_element",
-    srcs = [
-        "linear_simplex_element.cc",
-    ],
-    hdrs = [
-        "linear_simplex_element.h",
-    ],
-    deps = [
-        ":isoparametric_element",
-        "//common:default_scalars",
     ],
 )
 
@@ -700,9 +673,9 @@ drake_cc_googletest(
     deps = [
         ":dynamic_elasticity_element",
         ":linear_constitutive_model",
-        ":linear_simplex_element",
         "//common/test_utilities:eigen_matrix_compare",
         "//math:gradient",
+        "//multibody/fem:linear_simplex_element",
         "//multibody/fem:simplex_gaussian_quadrature",
     ],
 )
@@ -713,10 +686,10 @@ drake_cc_googletest(
         ":dynamic_elasticity_element",
         ":dynamic_elasticity_model",
         ":linear_constitutive_model",
-        ":linear_simplex_element",
         "//common/test_utilities:eigen_matrix_compare",
         "//geometry/proximity:make_box_mesh",
         "//math:gradient",
+        "//multibody/fem:linear_simplex_element",
         "//multibody/fem:simplex_gaussian_quadrature",
     ],
 )
@@ -736,10 +709,10 @@ drake_cc_googletest(
     deps = [
         ":elasticity_element",
         ":linear_constitutive_model",
-        ":linear_simplex_element",
         "//common/test_utilities:eigen_matrix_compare",
         "//math:geometric_transform",
         "//math:gradient",
+        "//multibody/fem:linear_simplex_element",
         "//multibody/fem:simplex_gaussian_quadrature",
     ],
 )
@@ -758,7 +731,6 @@ drake_cc_googletest(
         ":eigen_conjugate_gradient_solver",
         ":fem_solver",
         ":linear_constitutive_model",
-        ":linear_simplex_element",
         ":static_elasticity_element",
         ":static_elasticity_model",
         ":zeroth_order_state_updater",
@@ -766,6 +738,7 @@ drake_cc_googletest(
         "//common/test_utilities:expect_throws_message",
         "//geometry/proximity:make_box_mesh",
         "//math:gradient",
+        "//multibody/fem:linear_simplex_element",
         "//multibody/fem:simplex_gaussian_quadrature",
     ],
 )
@@ -794,27 +767,9 @@ drake_cc_googletest(
 )
 
 drake_cc_googletest(
-    name = "isoparametric_element_test",
-    deps = [
-        ":isoparametric_element",
-        ":linear_simplex_element",
-        "//common/test_utilities:eigen_matrix_compare",
-        "//common/test_utilities:expect_throws_message",
-    ],
-)
-
-drake_cc_googletest(
     name = "linear_constitutive_model_data_test",
     deps = [
         ":linear_constitutive_model_data",
-        "//common/test_utilities:eigen_matrix_compare",
-    ],
-)
-
-drake_cc_googletest(
-    name = "linear_simplex_element_test",
-    deps = [
-        ":linear_simplex_element",
         "//common/test_utilities:eigen_matrix_compare",
     ],
 )
@@ -868,11 +823,11 @@ drake_cc_googletest(
     name = "static_elasticity_element_test",
     deps = [
         ":linear_constitutive_model",
-        ":linear_simplex_element",
         ":static_elasticity_element",
         "//common/test_utilities:eigen_matrix_compare",
         "//common/test_utilities:expect_throws_message",
         "//math:gradient",
+        "//multibody/fem:linear_simplex_element",
         "//multibody/fem:simplex_gaussian_quadrature",
     ],
 )
@@ -881,12 +836,12 @@ drake_cc_googletest(
     name = "static_elasticity_model_test",
     deps = [
         ":linear_constitutive_model",
-        ":linear_simplex_element",
         ":static_elasticity_element",
         ":static_elasticity_model",
         "//common/test_utilities:eigen_matrix_compare",
         "//geometry/proximity:make_box_mesh",
         "//math:gradient",
+        "//multibody/fem:linear_simplex_element",
         "//multibody/fem:simplex_gaussian_quadrature",
     ],
 )
@@ -896,12 +851,12 @@ drake_cc_googletest(
     deps = [
         ":fem_solver",
         ":linear_constitutive_model",
-        ":linear_simplex_element",
         ":mesh_utilities",
         ":static_elasticity_element",
         ":static_elasticity_model",
         "//common/test_utilities:eigen_matrix_compare",
         "//geometry/proximity:make_box_mesh",
+        "//multibody/fem:linear_simplex_element",
         "//multibody/fem:simplex_gaussian_quadrature",
     ],
 )
