--- multibody/fixed_fem/dev/isoparametric_element.cc
@@ -1,1 +1,0 @@
-#include "drake/multibody/fixed_fem/dev/isoparametric_element.h"
